xf86drm: Handle NULL in drmCopyVersion
Just as it is already handled in the caller, `drmGetVersion`.

I'm not sure what the offending driver is, but the Ubuntu incidents
seem to be coming from a dual Intel/Nvidia machine. And they show
it is `card1` so I'm guessing `nvidia-drm` is the offender.

Bug-Ubuntu: https://bugs.launchpad.net/bugs/2104352

Gbp-Pq: Name xf86drm-Handle-NULL-in-drmCopyVersion.patch